+++ packages/cli/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @varlinor/cli
 
+## 2.0.1
+
+### Patch Changes
+
+- Updated dependencies
+  - @varlinor/node-tools@2.0.1
+
 ## 2.0.0
 
 ### Major Changes